layers: Cleanup util file
