+++ scripts/generators/object_tracker_generator.py
@@ -695,8 +695,8 @@ bool ObjectLifetimes::ReportUndestroyedDeviceObjects(VkDevice device, const Loca
                 # Do not generate validation code for the function's dispatchable parameter (the first one).
                 # Validation of such parameters is always successful based on the model of how the chassis
                 # dispactches the calls to ValidationObject. If invalid handle is used it will cause
-                # crash/corruption on the chassis level (in get_dispatch_key or later). And if correct handle
-                # is passed, then due to the mapping done by get_dispatch_key() the handle will belong to the
+                # crash/corruption on the chassis level (in GetDispatchKey or later). And if correct handle
+                # is passed, then due to the mapping done by GetDispatchKey() the handle will belong to the
                 # retrieved validation object, which guarantees positive result of the parenting test.
                 function_dispatchable_parameter = not is_struct and member == members[0]
                 if function_dispatchable_parameter:
@@ -749,7 +749,7 @@ bool ObjectLifetimes::ReportUndestroyedDeviceObjects(VkDevice device, const Loca
                     location = f'{errorLoc}.dot(Field::{member.name})'
                     if self.vk.commands[topCommand].device and self.vk.handles[member.type].instance:
                         # Use case when for device-level API call we should use instance-level validation object
-                        pre_call_validate += 'auto instance_data = GetLayerDataPtr(get_dispatch_key(instance), layer_data_map);\n'
+                        pre_call_validate += 'auto instance_data = GetLayerDataPtr(GetDispatchKey(instance), layer_data_map);\n'
                         pre_call_validate += 'auto instance_object_lifetimes = instance_data->GetValidationObject<ObjectLifetimes>();\n'
                         pre_call_validate += f'skip |= instance_object_lifetimes->ValidateObject({prefix}{member.name}, kVulkanObjectType{member.type[2:]}, {nullAllowed}, {param_vuid}, {parent_vuid}, {location}{parent_object_type});\n'
                     else:
